Experimental Change to Mood Detectinon
Instead of Averaging Values for Mood, it's instead modified to use MAJORITY VOTING to detect mood change:
-Files changed: Backend-> mood_detector.py

It takes about 25 seconds to change moods so instead of using the avgs. of moods detected it detects the 5 frames it captured what mood currently  displayed and the mood that appeared the most on those 5 frames triggers a mood change. Therefore mood change by majority voting